misc: Fix unknown-command help to include healthcheck
